tv-screener/pyproject.toml
[tool.poetry]
name = "tv-screener"
version = "1.0.0"
description = "TradingView Screener OpenAPI 3.1.0 generator"
authors = ["OpenAI <example@openai.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
orjson = "^3.10.7"
requests = "^2.32.3"
requests-cache = "^1.2.1"
tenacity = "^9.0.0"
openapi-spec-validator = "^0.7.1"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"